+++ src/triangle_group.cpp
@@ -29,14 +29,14 @@ G::G(void)
 {
 }
 
-G::G(const vector<int> reduction)
+G::G(const vector<long> reduction)
 {
   this->reduction = reduction;
   this->word = "";
 
-  for (int i = 0; i < 3; i++)
+  for (long i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (long j = 0; j < 3; j++)
     {
       {
         this->mat[i][j] = Ring(reduction);
@@ -49,7 +49,7 @@ void G::identity(void)
 {
   this->word="";
 
-  vector<int> c(this->reduction.size(), 0);
+  vector<long> c(this->reduction.size(), 0);
 
   Ring zero = Ring(c, this->reduction);
   
@@ -58,9 +58,9 @@ void G::identity(void)
   Ring id = Ring(c,this->reduction);
   
 
-  for (int i = 0; i < 3; i++)
+  for (long i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (long j = 0; j < 3; j++)
     {
       if(i==j)
       {
@@ -74,15 +74,15 @@ void G::identity(void)
   }
 }
 
-G G::operator%(const int &m) const
+G G::operator%(const long &m) const
 {
   G result = G(this->reduction);
 
   result.word = this->word;
 
-  for (int i = 0; i < 3; i++)
+  for (long i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (long j = 0; j < 3; j++)
     {
       result.mat[i][j] = this->mat[i][j] % m;
     }
@@ -96,9 +96,9 @@ void G::operator=(const G &other)
   this->word = other.word;
   this->reduction = other.reduction;
 
-  for (int i = 0; i < 3; i++)
+  for (long i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (long j = 0; j < 3; j++)
     {
       this->mat[i][j] = other.mat[i][j];
     }
@@ -112,11 +112,11 @@ G G::operator*(const G &other) const
   result.word = this->word + other.word;
   // result.reduction = other.reduction;
 
-  for (int i = 0; i < 3; i++)
+  for (long i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (long j = 0; j < 3; j++)
     {
-      for (int k = 0; k < 3; k++)
+      for (long k = 0; k < 3; k++)
       {
 	      result.mat[i][j] = result.mat[i][j] + this->mat[i][k] * other.mat[k][j];
       }
@@ -131,9 +131,9 @@ bool G::operator==(const G &other) const
 
   bool equal = true;
 
-  for (int i = 0; i < 3; i++)
+  for (long i = 0; i < 3; i++)
   {
-    for (int j = 0; j < 3; j++)
+    for (long j = 0; j < 3; j++)
     {
       equal &= (this->mat[i][j] == other.mat[i][j]);
 
@@ -151,9 +151,9 @@ string G::repr(void) const
 {
   string rep;
 
-  for (int i=0; i<3; i++)
+  for (long i=0; i<3; i++)
   {
-    for (int j=0; j<3; j++)
+    for (long j=0; j<3; j++)
     {
       rep += this->mat[i][j].repr() + "  ";
     }
@@ -198,16 +198,16 @@ size_t GHash::operator()(const G &obj) const
 {
     // https://stackoverflow.com/a/2595226/7236657
 
-    std::hash<int> hasher;
+    std::hash<long> hasher;
     size_t seed = 0;
 
-    vector<int> vals;
+    vector<long> vals;
 
-    for (int i = 0; i < 3; i++)
+    for (long i = 0; i < 3; i++)
     {
-      for(int j = 0; j < 3; j++)
+      for(long j = 0; j < 3; j++)
       {
-	for(int c: obj.mat[i][j].representation)
+	for(long c: obj.mat[i][j].representation)
 	{
           seed = hash_combine(seed, hasher(c));
 	}
@@ -220,11 +220,11 @@ size_t GHash::operator()(const G &obj) const
 
 
 // concatenate n copies of a word
-string word_power(const string &word, const int &n)
+string word_power(const string &word, const long &n)
 {
   string result("");
 
-  for (int i = 0; i < n; i++)
+  for (long i = 0; i < n; i++)
   {
     result += word;
   }
@@ -232,7 +232,7 @@ string word_power(const string &word, const int &n)
   return result;
 }
 
-TriangleGroup::TriangleGroup(const int &p, const int &q)
+TriangleGroup::TriangleGroup(const long &p, const long &q)
 {
   // Signature
   this->p = p;
@@ -279,17 +279,17 @@ void TriangleGroup::read_generators(void)
   ifstream file_Z(file_name_Z);
 
   string line;
-  int buffer;
-  vector<int> c;
+  long buffer;
+  vector<long> c;
 
   this->X = G(this->reduction);
   this->Y = G(this->reduction);
   this->Z = G(this->reduction);
 
   // 1D representation of the matrix; needs to be re-shaped at the end
-  vector<vector<int>> X_flat;
-  vector<vector<int>> Y_flat;
-  vector<vector<int>> Z_flat;
+  vector<vector<long>> X_flat;
+  vector<vector<long>> Y_flat;
+  vector<vector<long>> Z_flat;
 
   // -- Read in X -----------------------------------------------------
   if(file_X.is_open())
@@ -351,10 +351,10 @@ void TriangleGroup::read_generators(void)
   }
   file_Z.close();
 
-  int k=0;
-  for(int i=0; i<3; i++)
+  long k=0;
+  for(long i=0; i<3; i++)
   {
-    for(int j=0; j<3; j++)
+    for(long j=0; j<3; j++)
     {
       this->X.mat[i][j] = Ring(X_flat[k], this->reduction);
       this->Y.mat[i][j] = Ring(Y_flat[k], this->reduction);
